Fixed minor grammar problems and mis-translated
Okay. For the current version, we chinese can of, course, understand.
but any high-school students can find grammatical errors in it.
that's the first intention. some of the terms are not translated
We got to do it better.
Took time, checked the traditional chinese version and the original english.
Now you can see