An improved version of the "hack" section.
@mtt87 translated the idiomatic git verbs "merge" and "revert" to italian. His translation
is correct, but as far as I know Italian just say "fare un merge" o "fare revert" ("fare"
meaning "to do").

Also, the expression "e/o" (literally: and/or) is correct, but very formal. It can be found
in official legal documentation, but it's a bit weird in a web page.

I changed the original @mtt87's translation "hack" => "apportare modifiche" (literally:
"to produce modifications") with "sperimentare col codice" (literally: "to experiment/play
with code") which I find much more informal and colloquial.