Extract algolia docsearch config to layout front matter
This allows inheriting layouts to override the config as necessary.

The `algoliaOptions.facetFilters` property is kept as the base object into which the other properties will be merged. This is because the `lang/site` facets are presently being defined using the loadSearch params. Keeping them in the base allows existing layouts to not be changed.