Add style for diagrams
Companion commit to [1].

A few notes on the CSS:

`background-color: white` is necessary because draw.&#8203;io uses
that colour to create negative space (e. g. around labels) so the
background needs to match.

The `filter` and `-webkit-filter` properties are a workaround. They’re
needed because browsers don’t allow styling across document boundaries
(which an embedded `img` always constitutes).

The filters are applied *after* setting the background colour, which is
convenient because then the background colour still matches the negative
space inside the SVG.

[1]: https://github.com/Homebrew/brew/pull/8627