Use "en" as default lang
Post HTML renders with an empty `lang` attribute (i.e.,
`<html lang="">`) and this produces a validation warning. When a
`lang` value isn't provided in the front matter, we should either
use `en` as the default or omit the `lang` attribute. There's
something to be said for using an `en` default rather than omitting,
so this commit uses that approach.