Rename styles.scss to style.scss
We currently have a boilerplate 136 KB `style.css` file in our
generated output that presumably comes from something in the
`github-pages` gem. We don't use this file, so it's a waste of effort
to produce it but also we get validation errors for this file when we
check all the site files.

Renaming our `/assets/css/styles.scss` file to `style.scss` will
override the boilerplate file and effectively cause it not to be
produced. This path/filename is what GitHub suggests in documentation
about customizing Jekyll theme CSS, so this change aligns with that
guidance.