Remove trailing forward slash from void elements
It's not necessary to self-close a void element in HTML5 and this is
the cause of a lot of validation noise for us (unless restricted to
only errors). We currently have void elements that use a trailing
forward slash and some that don't, so this commit removes the forward
slash from related elements.

Unfortunately, our generated HTML will continue to include trailing
forward slashes in void elements from the `jekyll-feed` and
`jekyll-seo-tag` gems (where we can't control the output) but we can
at least address this in our templates.