Use SVG Homebrew logo for header
This reduces the page load size for this element from ~15 KiB to ~4.3 KiB,
saving around 10 KiB or reducing the total page load size by around 11%
as of the currently deployed master as reported by [this GTMetrix
report](https://gtmetrix.com/reports/brew.sh/qADTEmpE/).

The homebrew-256x256.png remains in the repo in case it's been
deeplinked somewhere but this header was its only usage in the codebase.