Replace self-closing tags in plugin output
The `jekyll-seo-tag` and `jekyll-feed` plugins use self-closing tags
(e.g., `<... />`) since they support XHTML but there's no way to
override this within the plugins. Self-closing void tags in HTML5 is
optional behavior but it creates a bunch of noise in HTML validators
unless we restrict the output to errors. Resolving this issue allows
us to optionally run an HTML validator in the default mode (which
also surfaces warning- and info-level issues) without having to sift
through hundreds of lines of output about self-closing tags.

This commit introduces some logic into the `base.html` layout that
removes trailing slashes in the plugin output, as this is the only
area where we have self-closing tags at this point.