Remove extra closing anchor tag
The `pagecontent.install.githubrelease` text already contains a
closing anchor tag, so the closing `a` tag in the `index.html`
template doesn't have a corresponding opening `a` tag. This currently
produces an HTML validation error: `error: Stray end tag “a”`.

This commit resolves the issue by removing the extra anchor tag in
the `index.html` template.