Add pretty permalink to alt lang index files
The existing alternative language files currently use an
`/index_{langcode}.html` path. While this is perfectly functional,
it's not as pretty as it could be. It's not uncommon for websites to
use a language path for alternative languages (e.g., `/fr/`) when
they use one domain for all languages.

This commit adds a `permalink` value to the front matter of all the
alternative language index files, so they will use cleaner URLs. This
also updates the URLs used in the `link` tags and the JavaScript for
the language `select` element accordingly.